EDU-370 Rename cdn controller and endpoints
