+++ src/server/controller-factory.js
@@ -1,5 +1,4 @@
 import { Container } from '../common/di.js';
-import CdnController from './cdn-controller.js';
 import RoomController from './room-controller.js';
 import I18nController from './i18n-controller.js';
 import UserController from './user-controller.js';
@@ -10,6 +9,7 @@ import StaticController from './static-controller.js';
 import ImportController from './import-controller.js';
 import ExportController from './export-controller.js';
 import LessonController from './lesson-controller.js';
+import StorageController from './storage-controller.js';
 import MySpaceController from './my-space-controller.js';
 import SettingsController from './settings-controller.js';
 import DocumentController from './document-controller.js';
@@ -25,7 +25,7 @@ const controllerTypes = [
   SearchController,
   UserController,
   DocumentController,
-  CdnController,
+  StorageController,
   ExportController,
   ImportController,
   ErrorController,
